requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile requirements.in
#
aiohttp==3.9.1
    # via
    #   -r requirements.in
    #   black
aiosignal==1.3.1
    # via aiohttp
apache-libcloud==3.8.0
    # via -r requirements.in
async-timeout==4.0.3
    # via aiohttp
asyncio==3.4.3
    # via -r requirements.in
asynctest==0.13.0
    # via -r requirements.in
attrs==23.2.0
    # via
    #   -r requirements.in
    #   aiohttp
    #   flake8-bugbear
autoflake==2.2.1
    # via -r requirements.in
azure-common==1.1.28
    # via
    #   azure-mgmt-commerce
    #   azure-mgmt-compute
    #   azure-mgmt-resource
azure-core==1.29.7
    # via
    #   -r requirements.in
    #   azure-identity
    #   azure-mgmt-core
    #   azure-storage-blob
    #   msrest
azure-identity==1.15.0
    # via -r requirements.in
azure-mgmt-commerce==6.0.0
    # via -r requirements.in
azure-mgmt-compute==32.0.0
    # via -r requirements.in
azure-mgmt-core==1.4.0
    # via
    #   azure-mgmt-commerce
    #   azure-mgmt-compute
    #   azure-mgmt-resource
azure-mgmt-resource==23.0.1
    # via -r requirements.in
azure-storage-blob==12.19.0
    # via -r requirements.in
backports-tarfile==1.2.0
    # via jaraco-context
bandit==1.7.6
    # via -r requirements.in
black==24.1a1
    # via -r requirements.in
build==1.0.3
    # via pip-tools
certifi==2023.11.17
    # via
    #   msrest
    #   requests
cffi==1.16.0
    # via cryptography
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via
    #   black
    #   pip-tools
    #   uvicorn
cryptography==42.0.0
    # via
    #   azure-identity
    #   azure-storage-blob
    #   msal
    #   pyjwt
    #   secretstorage
dataclasses==0.6
    # via -r requirements.in
docutils==0.21.2
    # via readme-renderer
fastapi==0.63.0
    # via -r requirements.in
flake8==6.1.0
    # via
    #   -r requirements.in
    #   flake8-bugbear
flake8-bugbear==23.12.2
    # via -r requirements.in
frozenlist==1.4.1
    # via
    #   aiohttp
    #   aiosignal
gitdb==4.0.11
    # via gitpython
gitpython==3.1.40
    # via bandit
h11==0.14.0
    # via uvicorn
idna==3.6
    # via
    #   requests
    #   yarl
importlib-metadata==8.4.0
    # via
    #   keyring
    #   twine
isodate==0.6.1
    # via
    #   azure-mgmt-compute
    #   azure-mgmt-resource
    #   azure-storage-blob
    #   msrest
isort==5.13.2
    # via -r requirements.in
jaraco-classes==3.4.0
    # via keyring
jaraco-context==6.0.1
    # via keyring
jaraco-functools==4.0.2
    # via keyring
jeepney==0.8.0
    # via
    #   keyring
    #   secretstorage
jinja2==3.1.2
    # via -r requirements.in
keyring==25.3.0
    # via twine
markdown-it-py==3.0.0
    # via rich
markupsafe==2.1.3
    # via jinja2
mccabe==0.7.0
    # via flake8
mdurl==0.1.2
    # via markdown-it-py
more-itertools==10.4.0
    # via
    #   jaraco-classes
    #   jaraco-functools
msal==1.26.0
    # via
    #   azure-identity
    #   msal-extensions
msal-extensions==1.1.0
    # via azure-identity
msrest==0.7.1
    # via azure-mgmt-commerce
multidict==6.0.4
    # via
    #   aiohttp
    #   yarl
mypy==1.8.0
    # via -r requirements.in
mypy-extensions==1.0.0
    # via
    #   black
    #   mypy
nh3==0.2.18
    # via readme-renderer
oauthlib==3.2.2
    # via requests-oauthlib
packaging==23.2
    # via
    #   black
    #   build
    #   msal-extensions
pathspec==0.12.1
    # via black
pbr==6.0.0
    # via stevedore
pip-tools==7.3.0
    # via -r requirements.in
pkginfo==1.10.0
    # via twine
platformdirs==4.1.0
    # via black
portalocker==2.8.2
    # via msal-extensions
pycodestyle==2.11.1
    # via flake8
pycparser==2.21
    # via cffi
pydantic==1.10.13
    # via
    #   -r requirements.in
    #   fastapi
pyflakes==3.1.0
    # via
    #   autoflake
    #   flake8
pygments==2.17.2
    # via
    #   readme-renderer
    #   rich
pyjwt[crypto]==2.8.0
    # via msal
pyproject-hooks==1.0.0
    # via build
python-dotenv==1.0.1
    # via -r requirements.in
pyyaml==6.0.1
    # via
    #   -r requirements.in
    #   bandit
readme-renderer==44.0
    # via twine
requests==2.31.0
    # via
    #   apache-libcloud
    #   azure-core
    #   msal
    #   msrest
    #   requests-oauthlib
    #   requests-toolbelt
    #   twine
requests-oauthlib==2.0.0
    # via msrest
requests-toolbelt==1.0.0
    # via twine
rfc3986==2.0.0
    # via twine
rich==13.7.0
    # via
    #   bandit
    #   twine
ruff==0.1.11
    # via -r requirements.in
secretstorage==3.3.3
    # via keyring
six==1.16.0
    # via
    #   azure-core
    #   isodate
smmap==5.0.1
    # via gitdb
starlette==0.13.6
    # via fastapi
stevedore==5.1.0
    # via bandit
tomli==2.0.1
    # via
    #   autoflake
    #   black
    #   build
    #   mypy
    #   pip-tools
    #   pyproject-hooks
twine==5.1.1
    # via -r requirements.in
typed-ast==1.5.4
    # via -r requirements.in
typing-extensions==4.9.0
    # via
    #   azure-core
    #   azure-mgmt-compute
    #   azure-storage-blob
    #   black
    #   mypy
    #   pydantic
    #   uvicorn
urllib3==2.1.0
    # via
    #   requests
    #   twine
uvicorn==0.25.0
    # via -r requirements.in
wheel==0.42.0
    # via
    #   -r requirements.in
    #   pip-tools
yarl==1.9.4
    # via aiohttp
zipp==3.20.1
    # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
